Renamed ctapipe/calib/camera/charge_extraction.py to ctapipe/calib/camera/charge_extractors.py
