+++ ctapipe/calib/camera/charge_extractors.py
