+++ ctapipe/calib/camera/tests/test_charge_extraction.py
@@ -4,7 +4,7 @@ from ctapipe.io import CameraGeometry
 import numpy as np
 from numpy.testing import assert_almost_equal
 
-from ctapipe.calib.camera.charge_extraction import FullIntegrator, \
+from ctapipe.calib.camera.charge_extractors import FullIntegrator, \
     SimpleIntegrator, GlobalPeakIntegrator, LocalPeakIntegrator, \
     NeighbourPeakIntegrator
